Split into centralized and decentralized storage for observers
